+++ CHANGELOG.md
@@ -57,6 +57,7 @@ Special thanks to the following contributors: @B1gG, @codac, @ezeholz, @khassel,
 - Fix calendar debug check
 - Really run prettier over all files
 - Fix logger.js after jest changes, use --forceExit running jest
+- Workaround for dev_console test using getWindowCount.
 
 ## [2.15.0] - 2021-04-01